requirements.txt
python-telegram-bot==20.7
apscheduler==3.10.4
python-dotenv==1.0.0
pytz==2024.1
aiosqlite==0.19.0